+++ src/components/docs-layout.js
@@ -1,33 +1,15 @@
-import React, { useState, createContext, useContext } from "react"
+import React, { useState } from "react"
 import Code from "./code"
 import { MDXProvider } from "@mdx-js/react"
 import { useStaticQuery, graphql, Link } from "gatsby"
 import { Caret, CaretDownWide } from "./icons"
+import RoutesService from "../lib/routes-service"
 
 const MAX_WIDTH = 1400
 const MAIN_WIDTH = 870
 const SIDEBAR_WIDTH = (MAX_WIDTH - MAIN_WIDTH) / 2
 
-const DocsRoutesContext = createContext()
-
-const NavSectionContext = createContext()
-
-function NavSection({ route, label, children }) {
-  return (
-    <li className="mb-8" key={route}>
-      <span className="text-gray-800 text-base+ font-medium">{label}</span>
-      <ul className="ml-2 mt-2 font-normal leading-snug">
-        <NavSectionContext.Provider value={route}>
-          {children}
-        </NavSectionContext.Provider>
-      </ul>
-    </li>
-  )
-}
-
-function NavLink(props) {
-  const sectionRoute = useContext(NavSectionContext)
-
+function NavLink({ fullPath, ...props }) {
   const isPartiallyActive = ({ isPartiallyCurrent }) => {
     return {
       className: isPartiallyCurrent
@@ -38,11 +20,7 @@ function NavLink(props) {
 
   return (
     <li className="py-1" key={props.route}>
-      <Link
-        getProps={isPartiallyActive}
-        to={`/docs/${sectionRoute}/${props.route}`}
-        {...props}
-      />
+      <Link getProps={isPartiallyActive} to={fullPath} {...props} />
     </li>
   )
 }
@@ -61,82 +39,67 @@ function MobileNavLink({ fullPath, ...otherProps }) {
   )
 }
 
-const docsRoutesService = {
-  _routes: [
-    {
-      name: "Documentation",
-      path: "/docs",
-      routes: [
-        {
-          path: "/getting-started",
-          name: "Getting started",
-          routes: [
-            { name: "Introduction", path: "/introduction" },
-            { name: "Installation", path: "/installation" },
-            { name: "Usage", path: "/usage" },
-          ],
-        },
-        {
-          path: "/examples",
-          name: "Examples",
-          routes: [
-            { name: "React", path: "/react" },
-            { name: "Vue", path: "/vue" },
-          ],
-        },
-      ],
-    },
-    {
-      name: "Examples",
-      path: "/examples",
-    },
-  ],
+function MobileNav({ routesService }) {
+  let [mobileSecondaryNavIsOpen, setMobileSecondaryNavIsOpen] = useState(false)
 
-  activePath: null,
-
-  // Transform _routes to include fullPath
-  get routes() {
-    function transformRoutes(routes = [], prefix = "") {
-      return routes.map(route => {
-        let fullPath = `${prefix}${route.path}`
-
-        route.fullPath = fullPath
-        route.routes = transformRoutes(route.routes, fullPath)
-
-        return route
-      }, [])
-    }
-
-    return transformRoutes(this._routes)
-  },
-
-  // Flatten all routes
-  get flattenedRoutes() {
-    function flatten(routes = []) {
-      return routes.reduce((flattenedRoutes, { name, fullPath, routes }) => {
-        return [...flattenedRoutes, { name, fullPath }, ...flatten(routes)]
-      }, [])
-    }
-
-    return flatten(this.routes)
-  },
-
-  // Return the active route
-  get activeRoute() {
-    return this.flattenedRoutes.find(route => {
-      return route.fullPath.match(this.activePath.replace(/\/+$/, ""))
-    })
-  },
-
-  // Return a subtree of routes under a path
-  routesForPath(path) {
-    return this.routes.find(route => route.path === "/docs").routes
-  },
+  return (
+    <div className="text-sm font-normal pl-5 py-1 text-gray-500 bg-gray-100 2xl:hidden">
+      <div className="flex items-center">
+        Documentation
+        <div
+          className={`flex items-center ${mobileSecondaryNavIsOpen ? "" : ""}`}
+        >
+          <Caret className="w-3 h-3 mx-2" />
+          <span>{routesService.activeRoute.name}</span>
+        </div>
+        <div className="ml-auto flex items-center">
+          <button
+            onClick={() =>
+              setMobileSecondaryNavIsOpen(!mobileSecondaryNavIsOpen)
+            }
+            className="flex items-center px-5 py-3 focus:outline-none"
+          >
+            <span
+              style={{
+                transform: mobileSecondaryNavIsOpen ? "rotate(180deg)" : "",
+              }}
+            >
+              <CaretDownWide className="w-4 h-4" />
+            </span>
+          </button>
+        </div>
+      </div>
+      {mobileSecondaryNavIsOpen && (
+        <div className="pt-1 pr-5">
+          <nav className="border-t border-gray-200 pt-5 pb-4 text-gray-700 text-base">
+            <ul className="pt-2w">
+              {routesService.routesForPath("/docs").map(route => (
+                <li className="mb-5" key={route.fullPath}>
+                  <div className="uppercase text-gray-400 text-sm font-medium">
+                    {route.name}
+                  </div>
+                  <ul>
+                    {route.routes.map(route => (
+                      <MobileNavLink
+                        fullPath={route.fullPath}
+                        key={route.fullPath}
+                        onClick={() => setMobileSecondaryNavIsOpen(false)}
+                      >
+                        {route.name}
+                      </MobileNavLink>
+                    ))}
+                  </ul>
+                </li>
+              ))}
+            </ul>
+          </nav>
+        </div>
+      )}
+    </div>
+  )
 }
 
 export default function DocsPage({ path, children }) {
-  let [mobileSecondaryNavIsOpen, setMobileSecondaryNavIsOpen] = useState(false)
-
   const data = useStaticQuery(graphql`
     query OnThisPageQuery {
       allMdx {
@@ -154,159 +117,93 @@ export default function DocsPage({ path, children }) {
   })
   let tableOfContentsItems = mdxPage && mdxPage.tableOfContents.items[0].items
 
-  docsRoutesService.activePath = path
+  let routesService = new RoutesService(path)
 
   return (
-    <DocsRoutesContext.Provider value={docsRoutesService}>
-      <div className="bg-white">
-        <div className="text-sm font-normal pl-5 py-1 text-gray-500 bg-gray-100 2xl:hidden">
-          <div className="flex items-center">
-            Documentation
-            <div
-              className={`flex items-center ${
-                mobileSecondaryNavIsOpen ? "" : ""
-              }`}
-            >
-              <Caret className="w-3 h-3 mx-2" />
-              <span>{docsRoutesService.activeRoute.name}</span>
-            </div>
-            <div className="ml-auto flex items-center">
-              <button
-                onClick={() =>
-                  setMobileSecondaryNavIsOpen(!mobileSecondaryNavIsOpen)
-                }
-                className="flex items-center px-5 py-3 focus:outline-none"
-              >
-                <span
-                  style={{
-                    transform: mobileSecondaryNavIsOpen ? "rotate(180deg)" : "",
-                  }}
-                >
-                  <CaretDownWide className="w-4 h-4" />
-                </span>
-              </button>
-            </div>
-          </div>
-          {mobileSecondaryNavIsOpen && (
-            <div className="pt-1 pr-5">
-              <nav className="border-t border-gray-200 pt-5 pb-4 text-gray-700 text-base">
-                <ul className="pt-2w">
-                  {docsRoutesService.routesForPath("/docs").map(route => (
-                    <li className="mb-5" key={route.fullPath}>
-                      <div className="uppercase text-gray-400 text-sm font-medium">
+    <div className="bg-white">
+      <MobileNav routesService={routesService} />
+
+      <div className="flex-1 flex">
+        <div
+          className="flex-shrink-0 bg-gray-100 border-r border-gray-200 hidden 2xl:block"
+          style={{
+            width: `calc(((100% - ${MAX_WIDTH}px)/ 2) + ${SIDEBAR_WIDTH}px)`,
+            paddingLeft: `calc((100% - ${MAX_WIDTH}px)/ 2)`,
+          }}
+        >
+          <nav className="pl-7 pt-14 pr-6 sticky top-0 leading-none h-screen overflow-y-scroll">
+            <ul className="mt-2">
+              {routesService.routesForPath("/docs").map(route => (
+                <li className="mb-8" key={route.fullPath}>
+                  <span className="text-gray-800 text-base+ font-medium">
+                    {route.name}
+                  </span>
+                  <ul className="ml-2 mt-2 font-normal leading-snug">
+                    {route.routes.map(route => (
+                      <NavLink fullPath={route.fullPath} key={route.fullPath}>
                         {route.name}
-                      </div>
-                      <ul>
-                        {route.routes.map(route => (
-                          <MobileNavLink
-                            fullPath={route.fullPath}
-                            key={route.fullPath}
-                            onClick={() => setMobileSecondaryNavIsOpen(false)}
-                          >
-                            {route.name}
-                          </MobileNavLink>
-                        ))}
-                      </ul>
-                    </li>
-                  ))}
-                </ul>
-              </nav>
-            </div>
-          )}
+                      </NavLink>
+                    ))}
+                  </ul>
+                </li>
+              ))}
+            </ul>
+          </nav>
         </div>
 
-        <div className="flex-1 flex">
-          <div
-            className="flex-shrink-0 bg-gray-100 border-r border-gray-200 hidden 2xl:block"
-            style={{
-              width: `calc(((100% - ${MAX_WIDTH}px)/ 2) + ${SIDEBAR_WIDTH}px)`,
-              paddingLeft: `calc((100% - ${MAX_WIDTH}px)/ 2)`,
-            }}
-          >
-            <nav className="pl-7 pt-14 pr-6 sticky top-0 leading-none h-screen overflow-y-scroll">
-              <ul className="mt-2">
-                <NavSection route="getting-started" label="Getting started">
-                  <NavLink route="introduction">Introduction</NavLink>
-                  <NavLink route="installation">Installation</NavLink>
-                  <NavLink route="usage">Usage</NavLink>
-                </NavSection>
-
-                <NavSection route="examples" label="Examples">
-                  <NavLink route="react">React</NavLink>
-                  <NavLink route="vue">Vue</NavLink>
-                </NavSection>
-
-                <NavSection route="api" label="API">
-                  <NavLink route="Association">Association</NavLink>
-                  <NavLink route="Collection">Collection</NavLink>
-                  <NavLink route="Db">Db</NavLink>
-                  <NavLink route="DbCollection">DbCollection</NavLink>
-                  <NavLink route="IdentityManager">IdentityManager</NavLink>
-                  <NavLink route="JSONAPISerializer">JSONAPISerializer</NavLink>
-                  <NavLink route="Model">Model</NavLink>
-                  <NavLink route="Response">Response</NavLink>
-                  <NavLink route="Schema">Schema</NavLink>
-                  <NavLink route="Serializer">Serializer</NavLink>
-                  <NavLink route="Server">Server</NavLink>
-                </NavSection>
-              </ul>
-            </nav>
-          </div>
-
-          <div
-            className="flex-1 max-w-full px-5 md:px-20 pt-7 md:pt-12 font-normal text-gray-700
+        <div
+          className="flex-1 max-w-full px-5 md:px-20 pt-7 md:pt-12 font-normal text-gray-700
           text-base leading-copy
           md:text-lg md:leading-relaxed"
-          >
-            <MDXProvider components={components}>{children}</MDXProvider>
-          </div>
+        >
+          <MDXProvider components={components}>{children}</MDXProvider>
+        </div>
 
-          <div
-            className="hidden 2xl:block flex-shrink-0"
-            style={{
-              width: `calc(((100% - ${MAX_WIDTH}px)/ 2) + ${SIDEBAR_WIDTH}px)`,
-              paddingRight: `calc((100% - 1408px)/ 2)`,
-            }}
-          >
-            <div className="pr-8">
-              <nav className="mt-32 ml-8 pl-6 sticky">
-                {tableOfContentsItems && (
-                  <>
-                    <p className="uppercase text-xs text-gray-800 font-medium tracking-wider">
-                      On this page
-                    </p>
+        <div
+          className="hidden 2xl:block flex-shrink-0"
+          style={{
+            width: `calc(((100% - ${MAX_WIDTH}px)/ 2) + ${SIDEBAR_WIDTH}px)`,
+            paddingRight: `calc((100% - 1408px)/ 2)`,
+          }}
+        >
+          <div className="pr-8">
+            <nav className="mt-32 ml-8 pl-6 sticky">
+              {tableOfContentsItems && (
+                <>
+                  <p className="uppercase text-xs text-gray-800 font-medium tracking-wider">
+                    On this page
+                  </p>
 
-                    <ul className="mt-2 font-normal text-sm">
-                      {tableOfContentsItems.map(item => (
-                        <li
-                          key={item.url}
-                          className="my-2 font-medium text-blue-500"
-                        >
-                          {item.title}
+                  <ul className="mt-2 font-normal text-sm">
+                    {tableOfContentsItems.map(item => (
+                      <li
+                        key={item.url}
+                        className="my-2 font-medium text-blue-500"
+                      >
+                        {item.title}
 
-                          {item.items && (
-                            <ul className="pl-4">
-                              {item.items.map(item => (
-                                <li
-                                  key={item.url}
-                                  className="my-2 font-medium text-blue-500"
-                                >
-                                  {item.title}
-                                </li>
-                              ))}
-                            </ul>
-                          )}
-                        </li>
-                      ))}
-                    </ul>
-                  </>
-                )}
-              </nav>
-            </div>
+                        {item.items && (
+                          <ul className="pl-4">
+                            {item.items.map(item => (
+                              <li
+                                key={item.url}
+                                className="my-2 font-medium text-blue-500"
+                              >
+                                {item.title}
+                              </li>
+                            ))}
+                          </ul>
+                        )}
+                      </li>
+                    ))}
+                  </ul>
+                </>
+              )}
+            </nav>
           </div>
         </div>
       </div>
-    </DocsRoutesContext.Provider>
+    </div>
   )
 }
